All persistence tests in one file
